rename folder login and change networks/api
