+++ src/networks/apis.js
@@ -1,8 +1,27 @@
 import axios from "axios";
 import CONST from "../utils/constants";
 
-const config = {
-  baseURL: CONST.BASE_API,
+const isDev = process.env.NODE_ENV === "development";
+
+const isLocalDev = (isDev) => {
+  let axiosConfig;
+  axiosConfig = axios.create();
+  if (isDev) {
+    const config = {
+      baseURL: CONST.BASE_API,
+    };
+    axiosConfig = axios.create(config);
+  }
+  return axiosConfig;
 };
 
-export const axiosInstance = axios.create(config);
+const axiosInstance = isLocalDev(isDev);
+axiosInstance.defaults.headers["Content-Type"] = "application/json";
+
+export default axiosInstance;
+
+// const config = {
+//   baseURL: CONST.BASE_API,
+// };
+
+// export const axiosInstance = axios.create(config);
