+++ src/routes/PrivateRoute.jsx
@@ -13,7 +13,7 @@ const useAuth = () => {
 
 const ProtectedRoutes = () => {
   const auth = useAuth();
-  return auth ? <Outlet /> : <Navigate to="/Login" />;
+  return auth ? <Outlet /> : <Navigate to="/login" />;
 };
 
 export default ProtectedRoutes;
